backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
moviepy==2.1.2
numpy==2.2.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
pillow==10.4.0
django-cors-headers==4.7.0